requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
azure-cosmos==4.7.0
python-dotenv==1.0.1
